Create Picture Cross Solver
*Work in Progress*

The objective of this program is determine the solution to any given Picture Cross puzzle. Picture cross is a game for smartphones that gives a puzzle with cells. The puzzle provides information for each row and column of the puzzle for how many squares of each row or column are selected. This objective of this code is to solve any picture cross puzzle once the user inputs all information the puzzle provides.